28. Generating Containers with connect() from React Redux (AddTodo)
Cause whe dont need the whole store in the AddTodo we only need the dispatch, so we read from the props. And in the mapDispatchToProps we return an object with the dispatch object. And like the AddTodo component is nor a container nor a presentational we change to let to wrap it an reasign.