03. Persisting the State to the Local Storage
We only save the todos part of the state so the filter is reset to
'SHOW_ALL' every time we reload and load the state from localstorage.

The node-uuid is for creating a unique identifier every time a todo is
added, cause without it whe we reload and read the localstorage and add
a new todo there is a warning cause the id is repeated cause we use a
local variable initialized at 0.

Now with throttle, even if the store gets updated really fast, we have a guarantee that we only write to localStorage once a second at most.